modifiedBy built based on production or test
The issue was that the profileUrl being used to set the person's uri in
the project update form is different between test and production. The
reason they are different is that in production it is
/vivo/display/<end-of-uri> whereas in test it is
/vivo/display?uri=<person's uri> which is good because it recognizes
that it is in test and not productio so builds the display uri
appropriately.

Because they are different we need a different approach to building the
person's uri. Unfortunately it isn't just available to be used here, so
we have to construct it.

We construct it given the subject (project) that this object (project
update) is being added to. That's why we modified the User.java class,
to make it available. Otherwise it's not available. Then we add on the
end of the display value, and this gives us the person's uri.