ServerInfo to get props from runtime.properties
ServerInfo had a lot of hard coded values in it, values that we need
to have configurable for the different test and dev environments.
Grabbing properties from the configuration file runtime.properties
required that the servlet context be available, since properties
can be based on the context. To avoid adding a lot of different
methods to ServerInfo based on the differetn request and session objects,
like you see in the configuration file class, the caller has to grab
the servlet context themselves. No new classes had to be created, only
classes that we've modified or created. And the context was always
available.