Added ckan api key to configuration
Had the ckan url but also needed the ckan api key. Also fixed problem where the return path was being set to the default namespace. But in the test
environment we need that to go to the test url instead of the production url. So created a baseNamespace to be the defaultNamespace but for the local url.