EditN3GeneratorVTwo wasn't checking for empty list
It was checking for a null list and if the elements in the list were null or empty, but wasn't checking to see if the list was empty.

Fixed a problem in addProjectUpdate.ftl where the group name for publications was set to publicationTitle but was expected to be publication.

Updated addProjectUpdate.ftl to make sure the first acSelection on the page is set to be empty

Probably don't need the code anywhere that checks for a file:/// but may as well keep it in there for the time being until we remove all of those objects
from the rdf.